chore: split GlobalFederationApiWithCache struct into separate file
